+++ src/impls/fsmultikey_map.rs
@@ -49,7 +49,7 @@ impl Builder {
             builder = builder.not_lazy();
         }
 
-        Ok(builder.try_build()?)
+        builder.try_build()
     }
 }
 
@@ -127,7 +127,7 @@ impl CidMap<Multikey> for FsMultikeyMap {
         let v = self.get(id)?;
 
         // get the paths
-        let (_, subfolder, file, lazy_deleted_file) = self.get_paths(&id)?;
+        let (_, subfolder, file, lazy_deleted_file) = self.get_paths(id)?;
 
         // remove the file if it exists
         if file.try_exists()? && file.is_file() {
@@ -143,11 +143,9 @@ impl CidMap<Multikey> for FsMultikeyMap {
         }
 
         // remove the subfolder if it is emtpy and we're not lazy
-        if subfolder.try_exists()? && subfolder.is_dir() {
-            if fs::read_dir(&subfolder)?.count() == 0 && !self.lazy {
-                fs::remove_dir(&subfolder)?;
-                debug!("fsmultikey_map: Removed subdir at: {}", subfolder.display());
-            }
+        if subfolder.try_exists()? && subfolder.is_dir() && fs::read_dir(&subfolder)?.count() == 0 && !self.lazy {
+            fs::remove_dir(&subfolder)?;
+            debug!("fsmultikey_map: Removed subdir at: {}", subfolder.display());
         }
 
         Ok(v)
